// src/style/configs/types/text.d.ts
export type TextVariantType =
  | "Heading/XLarge-36B"
  | "Heading/Large-28B"
  | "Heading/Medium-24B"
  | "Heading/Medium-24SB"
  | "Heading/Small-20B"
  | "Heading/Small-20SB"
  | "Subheading/XLarge-18B"
  | "Subheading/XLarge-18SB"
  | "Subheading/Large-16B"
  | "Subheading/Large-16SB"
  | "Subheading/Large-16M"
  | "Subheading/Medium-14B"
  | "Subheading/Medium-14SB"
  | "Subheading/Medium-14M"
  | "Subheading/Small-12B"
  | "Subheading/Small-12SB"
  | "Subheading/Small-12M"
  | "Subheading/XSmall-10B"
  | "Subheading/XSmall-10SB"
  | "Body/Large-16SB"
  | "Body/Large-16R"
  | "Body/Medium-14B"
  | "Body/Medium-14SB"
  | "Body/Medium-14R"
  | "Body/Small-12B"
  | "Body/Small-12SB"
  | "Body/Small-12R"
  | "Body/XSmall-10B"
  | "Body/XSmall-10R";